+++ src/hardware/drivers/motor/DShotMotor.cpp
@@ -1,7 +1,7 @@
 #include "DShotMotor.h"
 
 namespace MotorDrivers {
-    void DShotMotor::setOutput(int16_t output) {
+    void DShotMotor::setOutput(double output) {
         // TODO
     }
 }
