+++ src/hardware/drivers/motor/PwmMotor.h
@@ -13,7 +13,7 @@ namespace MotorDrivers {
     class PwmMotor : public Hardware::Motor {
     public:
         explicit PwmMotor(uint32_t pin);
-        void setOutput(int16_t output) override;
+        void setOutput(double output) override;
     private:
         ::Servo servo; // Without :: you get a conflict between the namespaced and global "Servo" classes
         uint32_t pin;
